pyproject.toml
[tool.poetry]
name = "internet-pet-shop"
version = "0.1.0"
description = ""
authors = ["Elena2004create <q62523474@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg = "^3.2.3"
pyqt6 = "^6.8.0"
qt6-tools = "^6.5.0.1.3"
fastapi = "^0.115.6"
gunicorn = "^23.0.0"
uvicorn = "^0.34.0"
pyjwt = "^2.10.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
